requirements.txt
fastapi~=0.110.1
uvicorn~=0.29.0
transformers~=4.39.3
llama_index~=0.10.27
llama_index_llms_llama_cpp~=0.1.3
llama_index_embeddings_huggingface~=0.2.0
llama_index_vector_stores_zep~=0.1.2
python_dotenv~=1.0.1
faiss_cpu~=1.8.0
docx2txt~=0.8